Implemented MakeTransfer functionality

+++ src/main/kotlin/com/dkb/bankingsystem/repositories/TransactionHistoryRepository.kt
@@ -3,9 +3,10 @@ package com.dkb.bankingsystem.repositories
 import com.dkb.bankingsystem.model.TransferHistory
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface TransactionHistoryRepository : JpaRepository<TransferHistory, Long> {
+interface TransactionHistoryRepository : JpaRepository<TransferHistory, Int> {
 
 
     fun findByFromAccount(fromAccount: String): List<TransferHistory>
+
 
 }
